Refactor QR code generation and redirection logic
This commit refactors the QR code generation and redirection process. It now uses the table ID instead of the table number for the QR code content. Refactors and improves the security config to allow the client app to access the API without JWT.
This change also uses a separate host for redirection.
Adds menu path to application properties file.